+++ app/src/main/java/com/example/academyhomework/di/ApplicationComponent.kt
@@ -3,16 +3,15 @@ package com.example.academyhomework.di
 
 import android.content.Context
 import com.example.academyhomework.MainActivity
+import com.example.academyhomework.RouterComponent
 import com.example.academyhomework.di.scopes.AppScope
 import com.example.academyhomework.presentation.playing_list.OnPlayingMovieFragment
 import com.example.academyhomework.presentation.details.MovieDetailsComponent
 import com.example.academyhomework.presentation.search.SearchFragment
 import com.example.academyhomework.services.db_update_work_manager.UpdateDBWorker
 import com.example.academyhomework.services.schedule_movie_work_manager.ScheduleNotificationWorker
-import dagger.Binds
 import dagger.BindsInstance
 import dagger.Component
-import dagger.Provides
 
 @AppScope
 @Component(modules = [RetrofitModule::class, AppModule::class, AppSubcomponents::class, RoomModule::class])
@@ -26,11 +25,12 @@ interface ApplicationComponent {
     //inject views
     fun inject(fragment: OnPlayingMovieFragment)
     fun inject(fragment: SearchFragment)
-    fun inject(act: MainActivity)
+
     //inject background
     fun inject(worker: UpdateDBWorker)
     fun inject(worker: ScheduleNotificationWorker)
 
     //subcomponents
     fun plusDetailsComponent(): MovieDetailsComponent.Factory
+    fun plusRouterComponent(): RouterComponent.Factory
 }
