+++ app/src/main/java/com/example/academyhomework/presentation/BaseFragment.kt
@@ -3,6 +3,9 @@ package com.example.academyhomework.presentation
 import androidx.fragment.app.Fragment
 import com.example.academyhomework.Router
 
+/**
+ *      Just marker yet
+ */
 abstract class BaseFragment: Fragment()
 {
      // val parentRouter: Router? get() = (activity as? Router)
