+++ app/src/main/res/layout/fragment_launch.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -16,15 +16,33 @@
         app:titleTextColor="@color/white_alpha_0.5"
         app:title="Movies@app"
         app:menu="@menu/search_menu"
+        tools:ignore="MissingConstraints" />
 
-        />
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/test_button"
+
+    <androidx.appcompat.widget.AppCompatCheckedTextView
+        android:id="@+id/tv_on_playing_more"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="/@\"
+        app:layout_constraintEnd_toEndOf="@id/tb_search"
+        app:layout_constraintTop_toTopOf="@id/rv_on_playing"
+        android:text="more"
+        android:textSize="20sp"
+        android:textColor="@color/link_color"
+        android:clickable="true"
         />
 
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_on_playing"
+            android:layout_width="match_parent"
+            android:layout_height="150dp"
+            android:orientation="horizontal"
+            app:layout_constraintTop_toBottomOf="@id/tb_search"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            tools:listitem="@layout/view_holder_movie_mini"
+            android:paddingTop="30dp"
+            />
 
-</FrameLayout>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
